refactor: 패키지 위치 변경 및 코드 리팩토링
