+++ src/main/java/com/sprint/mission/discodeit/auth/DiscodeitUserDetails.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service;
+package com.sprint.mission.discodeit.auth;
 
 import com.sprint.mission.discodeit.dto.User.UserDto;
 import java.util.Collection;
