+++ src/main/java/com/sprint/mission/discodeit/auth/DiscodeitUserDetailsService.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.service;
+package com.sprint.mission.discodeit.auth;
 
 import com.sprint.mission.discodeit.mapper.UserMapper;
 import com.sprint.mission.discodeit.repository.UserRepository;
