+++ src/main/java/com/sprint/mission/discodeit/common/init/AdminAccountInitializer.java
@@ -3,7 +3,6 @@ package com.sprint.mission.discodeit.common.init;
 import com.sprint.mission.discodeit.entity.Role;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
-import java.time.Instant;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
