+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,5 +1,6 @@
 package com.sprint.mission.discodeit.service;
 
+import com.sprint.mission.discodeit.auth.DiscodeitUserDetails;
 import com.sprint.mission.discodeit.dto.User.UserDto;
 
 public interface AuthService {
